Hookup Tab Buttons to Command
The two tab buttons call the same command and pass their text value,
which is bound to Days[0] or Days[1], to the SetDayCommand. The
SetDayCommand converts the string passed to it to an EventDay and
assigns it to the Day property. The Day property raises a
notification for the Event property. The Tab Button styles are
updated using nested multi bindings, BoolToStyleConverter and
MultiStringEqualityConverter, to reflect the day shown in the schedule.
The Events property returns either ThursdayEvents or FridayEvents.